% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBill.R
\name{getBill}
\alias{getBill}
\title{Return a specific bill}
\usage{
getBill(billID = NULL, legiKey = NULL)
}
\arguments{
\item{billID}{bill_id integer value}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
A list of bill details, including nested sponsors, history, texts,
roll calls, and session information.
}
\description{
Return a specific bill based on a legiscan bill_id
}
\examples{
\dontrun{
getBill(billID = 1633853)
}

}
