% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetList.R
\name{getDatasetList}
\alias{getDatasetList}
\title{Return list of available datasets}
\usage{
getDatasetList(state = NULL, year = NULL, legiKey = NULL)
}
\arguments{
\item{state}{US state 2 character abbreviation}

\item{year}{4 year digit}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
A data frame (tibble) of available datasets and their download
metadata. Columns come from the API response. An empty dataset list returns
zero rows and zero columns.
}
\description{
Return a list of available datasets, available to filter by state and year
}
\examples{
\dontrun{
getDatasetList(state = "TX", year = 2023)
}

}
