% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSponsoredList.R
\name{getSponsoredList}
\alias{getSponsoredList}
\title{Return df of bills an individual has sponsored}
\usage{
getSponsoredList(peopleID = NULL, legiKey = NULL)
}
\arguments{
\item{peopleID}{People id integer value}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
A data frame (tibble) of sponsored bills, excluding sponsor and
session metadata. Columns come from the API response. An empty bill list
returns zero rows and zero columns.
}
\description{
Returns a dataframe of bills that an indiviual legislator has sponsored, as specified using their people_id
}
\examples{
\dontrun{
getSponsoredList(peopleID = 5997)
}

}
