% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmendment.R
\name{getAmendment}
\alias{getAmendment}
\title{Return amendment document}
\usage{
getAmendment(amendmentID = NULL, legiKey = NULL)
}
\arguments{
\item{amendmentID}{amendment_id integer value from bill object}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Amendment document with metadata and base64 encoded document
}
\description{
Return amendment document in base64 encoded format based on an amendment_id
}
\examples{
getAmendment(amendmentID = 1234567)

}
