% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBill.R
\name{getBill}
\alias{getBill}
\title{Return a specific bill}
\usage{
getBill(billID = NULL, legiKey = NULL)
}
\arguments{
\item{billID}{bill_id integer value}
}
\description{
Return a specific bill based on a legiscan bill_id
}
\examples{
getBill(billID = 1633853)

}
