% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{Download dataset archive}
\usage{
getDataset(sessionID = NULL, accessKey = NULL, legiKey = NULL)
}
\arguments{
\item{sessionID}{session_id integer value (use session_id from getDatasetList)}

\item{accessKey}{access_key string value (use access_key from getDatasetList)}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Dataset archive with metadata and base64 encoded ZIP file
}
\description{
Return a base64 encoded ZIP archive of a specific dataset based on session_id and access_key
}
\examples{
\dontrun{
getDataset(sessionID = 1234, accessKey = "abc123def456")
}

}
