% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMasterlist.R
\name{getMasterList}
\alias{getMasterList}
\title{Return legiscan Master List}
\usage{
getMasterList(session = NULL, state = NULL, legiKey = NULL)
}
\arguments{
\item{session}{Session id integer value. Can be found with `getSessions`}

\item{state}{US state abbreviation}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Master List in dataframe format
}
\description{
Returns a legiscan Master List for a specified session or the most recent regular session if only state is provided
}
\examples{
getMasterList(session = 2108)
getMasterList(state = "TX")

}
