% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMasterListRaw.R
\name{getMasterListRaw}
\alias{getMasterListRaw}
\title{Return legiscan Master List for change detection}
\usage{
getMasterListRaw(session = NULL, state = NULL, legiKey = NULL)
}
\arguments{
\item{session}{Session id integer value. Can be found with `getSessions`}

\item{state}{US state abbreviation}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Master List of bill_id and change_hash in dataframe format
}
\description{
Returns a Master List of bill_id and change_hash values for a specified session
or the most recent regular session if only state is provided. Optimized for
detecting which bills have changed and need updating via `getBill`.
}
\examples{
\dontrun{
getMasterListRaw(session = 2108)
getMasterListRaw(state = "TX")
}

}
