% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitorList.R
\name{getMonitorList}
\alias{getMonitorList}
\title{Return associated monitor list}
\usage{
getMonitorList(record = "current", legiKey = NULL)
}
\arguments{
\item{record}{Record filter: "current" or "archived", or an exact year >= 2010}

\item{legiKey}{32 character API key from legiscan}
}
\value{
Monitored bills in dataframe format
}
\description{
Return items from the monitor list from the account associated with the legiKey provided
}
\examples{
\dontrun{
getMonitorList()
getMonitorList(record = "archived")
}

}
