% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitorList.R
\name{getMonitorList}
\alias{getMonitorList}
\title{Return monitored list}
\usage{
getMonitorList(legiKey = NULL, op = "getMonitorList")
}
\arguments{
\item{legiKey}{32 character API key from legiscan}
}
\description{
Return items from the monitored list from the account associated with legiKey provided
}
