% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitorList.R
\name{getMonitorList}
\alias{getMonitorList}
\title{Return associated monitor list}
\usage{
getMonitorList(legiKey = NULL)
}
\arguments{
\item{legiKey}{32 character API key from legiscan}
}
\description{
Return items from the monitor list from the account associated with the legiKey provided
}
