% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitorListRaw.R
\name{getMonitorListRaw}
\alias{getMonitorListRaw}
\title{Return associated monitor list for change detection}
\usage{
getMonitorListRaw(record = "current", legiKey = NULL)
}
\arguments{
\item{record}{Record filter: "current" or "archived", or an exact year >= 2010}

\item{legiKey}{32 character API key from legiscan}
}
\value{
Monitored bills with bill_id and change_hash in dataframe format
}
\description{
Return bill_id and change_hash values from the monitor list of the account
associated with the legiKey provided. Optimized for detecting which bills
have changed and need updating via `getBill`.
}
\examples{
\dontrun{
getMonitorListRaw()
getMonitorListRaw(record = "archived")
}

}
