% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerson.R
\name{getPerson}
\alias{getPerson}
\title{Return record on Person}
\usage{
getPerson(peopleID = NULL, legiKey = NULL)
}
\arguments{
\item{peopleID}{integer value from legiscan}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
A data frame (tibble) containing the individual record. Columns
come from the API response.
}
\description{
Return an individual record with basic information
}
\examples{
\dontrun{
getPerson(peopleID = 5997)
}

}
