% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerson.R
\name{getPerson}
\alias{getPerson}
\title{Return record on Person}
\usage{
getPerson(PeopleID = NULL, legiKey = NULL)
}
\arguments{
\item{PeopleID}{integer value from legiscan}
}
\description{
Return an individual record with basic information
}
