% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRollCall.R
\name{getRollCall}
\alias{getRollCall}
\title{Voting information for provided Roll Call}
\usage{
getRollCall(rollCallID = NULL, legiKey = NULL)
}
\arguments{
\item{rollCallID}{Roll Call ID integer}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Summary of Vote and nested list of individual votes by people id
}
\description{
Return vote detail summary for provided Roll Call and a nested list of individual votes
}
\examples{
\dontrun{
getRollCall(rollCallID = 1361957)
}

}
