% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSessionPeople.R
\name{getSessionPeople}
\alias{getSessionPeople}
\title{People from specified session}
\usage{
getSessionPeople(session = NULL, legiKey = NULL)
}
\arguments{
\item{session}{Session ID. Can be found with `getSessions`}

\item{legiKey}{32 character string provided by legiscan}
}
\description{
Return a dataframe of people from the specified session
}
\examples{
getSessionPeople(session = 2160)

}
