% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSessions.R
\name{getSessions}
\alias{getSessions}
\title{Return legislative session ids}
\usage{
getSessions(state = NULL, legiKey = NULL)
}
\arguments{
\item{state}{US State abbreviation}

\item{legiKey}{32 character string provided by legiscan.com}
}
\value{
A dataframe of all state legislative session ids
}
\description{
Return a dataframe for all session ids of specified state in legiscan's database
}
\examples{
getSessions(state = "MA")

}
