% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSponsoredList.R
\name{getSponsoredList}
\alias{getSponsoredList}
\title{Return df of bills an individual has sponsored}
\usage{
getSponsoredList(peopleID = NULL, legiKey = NULL)
}
\arguments{
\item{peopleID}{People id integer value}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Bills sponsored by a specified individual
}
\description{
Returns a dataframe of bills that an indiviual legislator has sponsored, as specified using their people_id
}
\examples{
getSponsoredList(peopleID = 5997)

}
