% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupplement.R
\name{getSupplement}
\alias{getSupplement}
\title{Return supplement document}
\usage{
getSupplement(supplementID = NULL, legiKey = NULL)
}
\arguments{
\item{supplementID}{supplement_id integer value from bill object}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Supplement document with metadata and base64 encoded document
}
\description{
Return supplement document in base64 encoded format based on a supplement_id
}
\examples{
getSupplement(supplementID = 1234567)

}
