% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getText.R
\name{getText}
\alias{getText}
\title{Return bill text document}
\usage{
getText(textID = NULL, legiKey = NULL)
}
\arguments{
\item{textID}{text_id integer value from bill object}

\item{legiKey}{32 character string provided by legiscan}
}
\value{
Bill text document with metadata and base64 encoded document
}
\description{
Return bill text document in base64 encoded format based on a text_id
}
\examples{
getText(textID = 1234567)

}
