% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legiSearch.R
\name{legiSearch}
\alias{legiSearch}
\title{Return search results from legiscan's database}
\usage{
legiSearch(
  query = NULL,
  state = "ALL",
  year = 2,
  session = NULL,
  page = 1,
  legiKey = NULL,
  op = "getSearch"
)
}
\arguments{
\item{query}{Enter your text}

\item{state}{Search the entire nation by default with 'ALL' or specify state using letter abbreviations}

\item{year}{Should be an integer. 1=All, 2=Current, 3=Recent, 4=Prior, >1900=Exact Year}

\item{session}{Limit search to specific session with a session_id}

\item{page}{Default is set to return Page 1. Up to 2,000 results per page}
}
\description{
Return up to 2,000 results at a time from legiscan's database
}
