% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setlegiKey.R
\name{setlegiKey}
\alias{setlegiKey}
\title{Set a legiscan api key}
\usage{
setlegiKey(APIkey, install = FALSE, overwrite = FALSE)
}
\arguments{
\item{APIkey}{32 character string given to you by legiscan.com}

\item{install}{Store the key in your .Renviron file for use across sessions}

\item{overwrite}{Overwrite a previously installed key in your .Renviron file}
}
\value{
Your legiscan API key is set for data requests
}
\description{
Every request requires a legiscan api key to be set in order to complete a request.
}
