% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setlegiKey.R
\name{setlegiKey}
\alias{setlegiKey}
\title{Set a legiscan api key}
\usage{
setlegiKey(APIkey, install = FALSE, overwrite = FALSE)
}
\arguments{
\item{APIkey}{32 character string given to you by legiscan.com}
}
\value{
Your legiscan API key is set for data requests
}
\description{
@description
 Every request requires a legiscan api key to be set in order to complete a request.
}
