% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMonitor.R
\name{setMonitor}
\alias{setMonitor}
\title{Add or remove bills from the monitor list}
\usage{
setMonitor(billIDs = NULL, action = NULL, stance = "watch", legiKey = NULL)
}
\arguments{
\item{billIDs}{One or more bill_id integer values to operate on}

\item{action}{Action to take on the bill list: "monitor", "remove", or "set"}

\item{stance}{Position on the bill: "watch", "support", or "oppose"}

\item{legiKey}{32 character API key from legiscan}
}
\value{
Status of each bill_id and the action taken in dataframe format
}
\description{
Interact with GAITS to add or remove bills from the monitor list of the
account associated with the legiKey provided, or set a stance on monitored bills
}
\examples{
\dontrun{
setMonitor(billIDs = c(150334, 141690), action = "monitor")
setMonitor(billIDs = 1533377, action = "set", stance = "support")
}

}
